requirements.txt
Flask
Flask-MySQLdb
Werkzeug
PyMySQL
python-dotenv
stripe
Flask-Mail
celery
flask_migrate
